KInventory: Move from kq.h and itemmenu.cpp to inventory.cpp/.h
