+++ src/shopmenu.cpp
@@ -33,6 +33,7 @@
 #include "fade.h"
 #include "gfx.h"
 #include "input.h"
+#include "inventory.h"
 #include "itemdefs.h"
 #include "itemmenu.h"
 #include "kq.h"
